feat: split slider into a control and field composables
