+++ packages/core/src/useSlider/index.ts
@@ -1,2 +1,3 @@
 export * from './useSlider';
 export * from './useSliderThumb';
+export * from './useSliderControl';
